generator/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
bvhio==1.5.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
etils==1.13.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
mujoco==3.3.7
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
polyscope==2.5.0
protobuf==6.33.1
py-cpuinfo==9.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pyglm==2.8.2
PyOpenGL==3.1.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
sentry-sdk==2.45.0
six==1.17.0
smmap==5.0.2
spatial-transform==1.3.3
sympy==1.14.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
trimesh==4.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.23.0
zipp==3.23.0
